clean up components
