+++ app/javascript/dashboard/routes/dashboard/settings/integrations/Index.vue
@@ -8,39 +8,14 @@
             :key="item.id"
             class="small-12 columns integration"
           >
-            <div class="row">
-              <div class="integration--image">
-                <img :src="'/assets/dashboard/integrations/' + item.logo" />
-              </div>
-              <div class="column">
-                <h3 class="integration--title">
-                  {{ item.name }}
-                </h3>
-                <p class="integration--description">
-                  {{ item.description }}
-                </p>
-              </div>
-              <div class="small-2 column button-wrap">
-                <router-link
-                  :to="
-                    frontendURL(
-                      `accounts/${accountId}/settings/integrations/` + item.id
-                    )
-                  "
-                >
-                  <div v-if="item.enabled">
-                    <button class="button nice">
-                      {{ $t('INTEGRATION_SETTINGS.WEBHOOK.CONFIGURE') }}
-                    </button>
-                  </div>
-                </router-link>
-                <div v-if="!item.enabled">
-                  <a :href="item.action" class="button success nice">
-                    {{ $t('INTEGRATION_SETTINGS.CONNECT.BUTTON_TEXT') }}
-                  </a>
-                </div>
-              </div>
-            </div>
+            <integration
+              :integration-id="item.id"
+              :integration-logo="item.logo"
+              :integration-name="item.name"
+              :integration-description="item.description"
+              :integration-enabled="item.enabled"
+              :integration-action="item.action"
+            />
           </div>
         </div>
       </div>
@@ -49,24 +24,19 @@
 </template>
 <script>
 import { mapGetters } from 'vuex';
-import { frontendURL } from '../../../../helper/URLHelper';
-import globalConfigMixin from 'shared/mixins/globalConfigMixin';
+import Integration from './Integration';
 
 export default {
-  mixins: [globalConfigMixin],
+  components: {
+    Integration,
+  },
   computed: {
     ...mapGetters({
-      currentUser: 'getCurrentUser',
-      globalConfig: 'globalConfig/get',
-      accountId: 'getCurrentAccountId',
       integrationsList: 'integrations/getIntegrations',
     }),
   },
   mounted() {
     this.$store.dispatch('integrations/get');
-  },
-  methods: {
-    frontendURL,
   },
 };
 </script>
